código simplificado

+++ RequestsNFCE.py
@@ -2,10 +2,7 @@ import time, json, logging, os
 from selenium.webdriver.common.by import By
 from selenium.webdriver.support.ui import WebDriverWait
 from selenium.webdriver.support import expected_conditions as EC
-from util import (
-    capturar_horario, iniciar_navegador, aguardar_tempo_para_clicar, 
-    realizar_login, acessar_link, criar_lista_solicitacoes
-)
+from util import (capturar_data_hora, iniciar_navegador_firefox, autenticar_sefaz, acessar_pagina, montar_lista_solicitacoes)
 
 # Configuração de logs
 logging.basicConfig(
@@ -37,7 +34,7 @@ def salvar_solicitacoes(solicitacoes):
     with open(caminho_json, "w", encoding="utf-8") as arquivo:
         json.dump(solicitacoes, arquivo, indent=4, ensure_ascii=False)
 
-def preencher_datas(navegador, data_inicio, data_fim, espera=2):
+def inserir_datas_formulario(navegador, data_inicio, data_fim, espera=2):
     wait = WebDriverWait(navegador, espera)
     data_inicio_elemento = wait.until(EC.presence_of_element_located((By.XPATH, NFCE["XPATH_DATA_INICIO"])))
     data_inicio_elemento.clear()
@@ -56,61 +53,68 @@ def preencher_campo_iframe(navegador, ie_empresa, espera=2):
     botao_pesquisar.click()
     navegador.switch_to.default_content()
 
-def selecionar_xml_e_executar(navegador, espera=2):
+def espera_para_clicar():
+    now = time.localtime()
+    segundo = now.tm_sec
+    if segundo < 1: tempo_espera = 1
+    elif segundo < 31: tempo_espera = 31
+    else:
+        time.sleep(60 - segundo)
+        now = time.localtime()
+        segundo = now.tm_sec
+        tempo_espera = 1
+    time.sleep(max(0, tempo_espera - segundo))
+
+def selecionar_xml_executar(navegador, espera=2):
     wait = WebDriverWait(navegador, espera)
     dropdown_xml = wait.until(EC.presence_of_element_located((By.XPATH, NFCE["XPATH_DROPDOWN_XML"])))
     dropdown_xml.click()
     opcao_xml = wait.until(EC.presence_of_element_located((By.XPATH, NFCE["XPATH_OPCAO_XML"])))
     opcao_xml.click()
     botao_executar = wait.until(EC.element_to_be_clickable((By.XPATH, NFCE["XPATH_BOTAO_EXECUTAR"])))
-    aguardar_tempo_para_clicar()
+    espera_para_clicar()
     botao_executar.click()
 
-def solicitacoes_nfce(navegador, empresa, espera=2):
-    """Executa o processo de solicitação da NFC-e e atualiza os dados no JSON."""
-    try:
-        preencher_datas(navegador, empresa["data_ini"], empresa["data_fim"], espera)
-        preencher_campo_iframe(navegador, empresa["inscricao_estadual"], espera)
-        selecionar_xml_e_executar(navegador, espera)
 
-        # Captura o horário e atualiza no JSON
-        horario_capturado = capturar_horario()
+def solicitar_nfce(navegador, empresa, espera=2):
+    try:
+        inserir_datas_formulario(navegador, empresa["data_ini"], empresa["data_fim"], espera)
+        preencher_campo_iframe(navegador, empresa["inscricao_estadual"], espera)
+        selecionar_xml_executar(navegador, espera)
+        horario_capturado = capturar_data_hora()
         empresa["horario"] = horario_capturado
         empresa["solicitado"] = True
-
         time.sleep(10)
         return True
     except Exception as e:
         logging.error(f"Erro ao solicitar NFCE: {e}")
         return False
 
-def excluir_solicitacoes_anteriores():
-    """Remove o arquivo de solicitações, se existir."""
-    caminho_json = "json_files/solicitacoes.json"
-    if os.path.exists(caminho_json):
-        os.remove(caminho_json)
+def remover_solicitacoes_anteriores():
+    caminho_json = os.path.join(os.path.dirname(__file__), 'json_files/solicitacoes.json')
+    if os.path.exists(caminho_json): os.remove(caminho_json)
 
 def executar_processo_requests_nfce():
     """Executa todo o processo de solicitação de NFC-e usando os dados do JSON."""
-    excluir_solicitacoes_anteriores()
-    criar_lista_solicitacoes("NFCE")  # Gera novo JSON
+    remover_solicitacoes_anteriores()
+    montar_lista_solicitacoes("NFCE")  # Gera novo JSON
     solicitacoes = carregar_solicitacoes()  # Carrega o novo JSON
 
     navegador = None
     logging.info("Iniciando processo de solicitações de NFC-e...")
-    
+
     try:
-        navegador = iniciar_navegador()
-        if navegador and realizar_login(navegador):
+        navegador = iniciar_navegador_firefox()
+        if navegador and autenticar_sefaz(navegador):
             link = NFCE["LINK_SEFAZ_NFCE"]
-            acessar_link(navegador, link)
+            acessar_pagina(navegador, link)
 
             for empresa in solicitacoes:
                 if not empresa["solicitado"]:
-                    if solicitacoes_nfce(navegador, empresa):
+                    if solicitar_nfce(navegador, empresa):
                         empresa["solicitado"] = True  # Marca como solicitado
                         salvar_solicitacoes(solicitacoes)  # Salva atualização no JSON
-                    acessar_link(navegador, link)
+                    acessar_pagina(navegador, link)
                 else:
                     logging.info(f"Já solicitado: IE {empresa['inscricao_estadual']}, Data Início {empresa['data_ini']}, Data Fim {empresa['data_fim']}")
